revised code to smaller size
